Removes findIds method.
This method can't exist at this level because not only does it enforce a an implied pattern for caching, but it also crosses two different interfaces - identifiable items, and "where" based datastores.

This likely can exist in traits, but should not be something that is considered a part of the interfaces at this level.

We might consider re-introducing this in another interface if it's determined to be required somewhere.